// src/gen/sls/sls_pb.ts
export type LogJson = {
  /**
   * UNIX Time Format
   *
   * @generated from field: required uint32 Time = 1;
   */
  Time?: number;

  /**
   * @generated from field: repeated sls.Log.Content Contents = 2;
   */
  Contents?: Log_ContentJson[];

  /**
   * @generated from field: optional fixed32 TimeNs = 4;
   */
  TimeNs?: number;
};
// ...
export type Log_ContentJson = {
  /**
   * @generated from field: required string Key = 1;
   */
  Key?: string;

  /**
   * @generated from field: required string Value = 2;
   */
  Value?: string;
};
// ...
export type LogTagJson = {
  /**
   * @generated from field: required string Key = 1;
   */
  Key?: string;

  /**
   * @generated from field: required string Value = 2;
   */
  Value?: string;
};
// ...
export type LogGroupJson = {
  /**
   * @generated from field: repeated sls.Log Logs = 1;
   */
  Logs?: LogJson[];

  /**
   * reserved fields
   *
   * @generated from field: optional string Reserved = 2;
   */
  Reserved?: string;

  /**
   * @generated from field: optional string Topic = 3;
   */
  Topic?: string;

  /**
   * @generated from field: optional string Source = 4;
   */
  Source?: string;

  /**
   * @generated from field: repeated sls.LogTag LogTags = 6;
   */
  LogTags?: LogTagJson[];
};